Rename all files in src/javascripts
