+++ src/javascripts/ConsoleStyle.js
