+++ src/javascripts/MissionStyle.js
